編集時画像削除条件追加

+++ my_app/views/edit.py
@@ -70,7 +70,7 @@ def edit_result():
     if file.filename != '':
         filename = save_file(file, file.filename, UPLOAD_FOLDER)
     # 画像更新の際、以前のファイルを削除する
-    if session['user']['image_file_path'] != '':
+    if session['user']['image_file_path'] != '' and filename is not None:
         remove_file(session['user']['image_file_path'])
     employee_id = request.form['employee_id']
     sql = issue_sql('edit_check')
